Changes as requested in pr
