+++ desktop/src/renderer/src/components/transfer/lan-views/finish-view/RequestFailed.tsx
@@ -1,0 +1,9 @@
+import { Stack, Typography } from "@mui/material";
+
+export const RequestFailed = () => {
+  return (
+    <Stack gap={2}>
+      <Typography>An unexpected error occured.</Typography>
+    </Stack>
+  );
+};
