+++ desktop/src/renderer/src/components/transfer/lan-views/finish-view/RequestLoading.tsx
@@ -1,0 +1,25 @@
+import { LinearProgress, Stack, Typography } from "@mui/material";
+
+export const RequestLoading = () => {
+  return (
+    <Stack gap={2}>
+      <Typography variant="h3">Sending your records</Typography>
+      <LinearProgress
+        sx={{
+          "& .MuiLinearProgress-bar": {
+            backgroundColor: "var(--upload-progress-bar)",
+          },
+        }}
+      />
+      <Typography>
+        We’re processing your request, and sending the records to DATS. This
+        should only take about 30 seconds. Please do not close the application
+        or navigate away.
+      </Typography>
+      <Typography>
+        Once the loading is complete, a success confirmation message will appear
+        on this page.
+      </Typography>
+    </Stack>
+  );
+};
