+++ desktop/src/renderer/src/components/transfer/lan-views/finish-view/index.ts
@@ -1,0 +1,1 @@
+export * from "./LanFinishView";
