+++ desktop/src/renderer/src/components/transfer/lan-views/index.ts
@@ -1,5 +1,5 @@
 export * from "./LanUploadFileListView";
 export * from "./LanUploadTransferFormView";
 export * from "./LanSubmissionAgreementView";
-export * from "./LanFinishView";
+export * from "./finish-view";
 export * from "./LanConfirmationView";
