+++ desktop/src/renderer/src/pages/transfers/LanTransfer.tsx
@@ -25,7 +25,7 @@ type Folder = {
   metadataProgress: number;
 };
 
-type Change = {
+type FolderUploadChange = {
   originalFolderPath: string;
   newFolderPath?: string;
   deleted: boolean;
@@ -60,7 +60,7 @@ export const LanTransferPage = ({ accessToken }: { accessToken?: string }) => {
   >({});
   const [foldersToProcess, setFoldersToProcess] = useState<string[]>([]);
   const [folders, setFolders] = useState<Folder[]>([]);
-  const [changes, setChanges] = useState<Change[]>([]);
+  const [changes, setChanges] = useState<FolderUploadChange[]>([]);
 
   // Justify changes
   const [showJustifyChangesModal, setShowJustifyChangesModal] = useState(false);
@@ -84,7 +84,6 @@ export const LanTransferPage = ({ accessToken }: { accessToken?: string }) => {
   };
 
   // Handle metadata progress and completion events
-  // biome-ignore lint/correctness/useExhaustiveDependencies: <explanation>
   useEffect(() => {
     const handleProgress = (
       event: CustomEvent<{ source: string; progressPercentage: number }>
@@ -171,7 +170,6 @@ export const LanTransferPage = ({ accessToken }: { accessToken?: string }) => {
   }, []);
 
   // Handle buffer progress and completion events
-  // biome-ignore lint/correctness/useExhaustiveDependencies: <explanation>
   useEffect(() => {
     const handleProgress = (
       event: CustomEvent<{ source: string; progressPercentage: number }>
@@ -260,7 +258,6 @@ export const LanTransferPage = ({ accessToken }: { accessToken?: string }) => {
   }, []);
 
   // Get folder metadata and buffers after file list uploaded
-  // biome-ignore lint/correctness/useExhaustiveDependencies: <explanation>
   useEffect(() => {
     if (foldersToProcess.length > 0) {
       const pathsToProcess = [...foldersToProcess];
@@ -303,7 +300,6 @@ export const LanTransferPage = ({ accessToken }: { accessToken?: string }) => {
     }
   }, [foldersToProcess]);
 
-  // biome-ignore lint/correctness/useExhaustiveDependencies: <explanation>
   useEffect(() => {
     if (currentViewIndex === 3) {
       // Open of upload view
@@ -498,7 +494,6 @@ export const LanTransferPage = ({ accessToken }: { accessToken?: string }) => {
     return; // Return so linting doesnt complain about some paths not returning.
   };
 
-  // biome-ignore lint/correctness/useExhaustiveDependencies: <explanation>
   useEffect(() => {
     if (!fileList) {
       setFolders([]);
@@ -507,7 +502,6 @@ export const LanTransferPage = ({ accessToken }: { accessToken?: string }) => {
     parseFileList();
   }, [fileList]);
 
-  // biome-ignore lint/correctness/useExhaustiveDependencies: <explanation>
   useEffect(() => {
     if (currentViewIndex === 3) {
       // Open of upload view
